votum-bot/requirements.txt
discord.py>=2.3.2,<3.0
aiosqlite>=0.19.0
fastapi>=0.103.0
uvicorn[standard]>=0.23.0
jinja2>=3.1.2
python-dotenv>=1.0.0
pydantic>=2.3.0
websockets>=11.0.3
itsdangerous>=2.1.2  # for CSRF tokens
PyJWT>=2.8.0
requests-oauthlib>=1.3.1
python-multipart>=0.0.6
starlette>=0.27.0